+++ apps/frontend/src/libs/components/protected-route/protected-route.tsx
@@ -3,6 +3,8 @@ import { type AppRoute } from "~/libs/enums/enums.js";
 import { useAppSelector } from "~/libs/hooks/hooks.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
+import { AuthWrapper } from "../auth-wrapper/auth-wrapper.js";
+
 type Properties = {
 	component: React.ReactNode;
 	redirectTo: ValueOf<typeof AppRoute>;
@@ -15,7 +17,7 @@ const ProtectedRoute: React.FC<Properties> = ({
 	const user = useAppSelector(({ auth }) => auth.user);
 
 	if (user) {
-		return component;
+		return <AuthWrapper>{component}</AuthWrapper>;
 	}
 
 	return <Navigate replace to={redirectTo} />;