+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -1,3 +1,3 @@
-type IconName = "wheelActive" | "wheelInactive";
+type IconName = "crossedEye" | "eye" | "wheelActive" | "wheelInactive";
 
 export { type IconName };